backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.3
pydantic-settings>=2.0.3
sqlalchemy>=2.0.18
asyncpg>=0.28.0
alembic>=1.11.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
httpx>=0.24.1
redis>=4.6.0
